Fail if DAPPER_OUTPUT is empty
This is checked explicitly with -z, but shellcheck still suggests
protecting the rm with ${DAPPER_OUTPUT:?} (which causes the shell to
exit if DAPPER_OUTPUT is empty), and this seems like a safer solution
than disabling the shellcheck check.

+++ scripts/shared/cleanup.sh
@@ -39,7 +39,7 @@ run_if_defined pre_cleanup
 
 # run_parallel expects cluster names as a single argument
 run_parallel "${clusters[*]}" delete_cluster
-[[ -z "${DAPPER_OUTPUT}" ]] || rm -rf "${DAPPER_OUTPUT}"/*
+[[ -z "${DAPPER_OUTPUT}" ]] || rm -rf "${DAPPER_OUTPUT:?}"/*
 
 stop_local_registry
 docker system prune --volumes -f
